# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus Hosting API
  version: 1.0.0
  description: |
    Production-grade federated static site hosting network.
    Every node exposes this API. Authentication uses session cookies (browser)
    or Bearer API tokens (CLI / CI). Inter-node requests use Ed25519 signatures.
  contact:
    name: The No Hands Company
    url: https://github.com/The-No-Hands-company/Nexus-Hosting
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: Local node API
...
paths:
...
  /tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: List active API tokens for the current user
      security:
        - sessionAuth: []
        - bearerToken: []
      responses:
        "200":
          description: List of tokens (hashes never returned)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createToken
      tags: [tokens]
      summary: Create a new API token (plaintext returned once)
      security:
        - sessionAuth: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenBody"
      responses:
        "201":
          description: Token created — plaintext in `token` field, shown once only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /tokens/{id}:
    delete:
      operationId: revokeToken
      tags: [tokens]
      summary: Revoke an API token
      security:
        - sessionAuth: []
        - bearerToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Token revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Nodes ───────────────────────────────────────────────────────────────────
...
  /sites:
    get:
      operationId: listSites
      tags: [sites]
      summary: List all sites (paginated, filterable)
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, migrating]
        - name: ownerId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated site list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSites"
    post:
      operationId: createSite
      tags: [sites]
      summary: Register a new site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteBody"
      responses:
        "201":
          description: Site registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /sites/{id}:
    get:
      operationId: getSite
      tags: [sites]
      summary: Get site by ID
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Site details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateSite
      tags: [sites]
      summary: Update site metadata
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSiteBody"
      responses:
        "200":
          description: Updated site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Delete a site and all its files
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Site deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Deployment ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: sid
      description: Browser session cookie set by /login flow
    bearerToken:
      type: http
      scheme: bearer
      description: API token created via POST /tokens (prefix fh_)
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
        code:
          type: string
        status:
          type: integer
    PaginationMeta:
      type: object
      required: [total, page, limit, totalPages, hasNextPage, hasPrevPage]
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPrevPage:
          type: boolean
...
    Site:
      type: object
      required: [id, name, domain, status, siteType, ownerName, ownerEmail, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        domain:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, suspended, migrating]
        siteType:
          type: string
          enum: [static, dynamic, blog, portfolio, other]
        visibility:
          type: string
          enum: [public, private, password]
        ownerName:
          type: string
        ownerEmail:
          type: string
        ownerId:
          type: ["string", "null"]
        primaryNodeId:
          type: ["integer", "null"]
        primaryNodeDomain:
          type: ["string", "null"]
        replicaCount:
          type: integer
        storageUsedMb:
          type: number
        monthlyBandwidthGb:
          type: number
        hitCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    PaginatedSites:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Site"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    CreateSiteBody:
      type: object
      required: [name, domain, siteType, ownerName, ownerEmail]
      properties:
        name:
          type: string
        domain:
          type: string
        description:
          type: string
        siteType:
          type: string
          enum: [static, dynamic, blog, portfolio, other]
        ownerName:
          type: string
        ownerEmail:
          type: string
        primaryNodeId:
          type: integer
    UpdateSiteBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, suspended, migrating]
        siteType:
          type: string
          enum: [static, dynamic, blog, portfolio, other, nlpl]
        primaryNodeId:
          type: integer
        spaRouting:
          type: integer
          enum: [0, 1]
          description: "1 = serve index.html for unknown paths (SPA), 0 = strict 404 (MPA)"
...
    ApiToken:
      type: object
      required: [id, name, tokenPrefix, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        tokenPrefix:
          type: string
        lastUsedAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        revokedAt:
          type: ["string", "null"]
    CreatedToken:
      allOf:
        - $ref: "#/components/schemas/ApiToken"
        - type: object
          required: [token]
          properties:
            token:
              type: string
              description: "Plaintext token — shown once only. Store securely."
    CreateTokenBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        expiresInDays:
          type: integer
          minimum: 1
          maximum: 365
...